Made playbackbot top level robot
